requirements.txt
Flask==3.0.2
gspread==6.0.2
google-auth==2.28.1
toml==0.10.2
gunicorn==21.2.0
requests==2.31.0
yfinance==0.2.37
pandas==2.2.1